Fix Rollupstats deadlock by forcing MRTask to compute Rollups upfront. Repro: DeepLearningProstateTest on 3 nodes on previous commit.
Rollups were called from inside a map(), which led to too many RPCs
(nodes*threads) asking for Rollups to be computed from the home node, and the
thread priority of the waiting RPCs were at the same level as the priority of
the thread initiating the actual Rollup "Roll" MRTask, this exceeded the
threadpool's capacity.

This fix reduces the number of same-priority-threads-for-rollups, but another
fix would perform the rollup computation with a different threading mechanism.

Tomas knows more details, and intends to fix this.